requirements.txt
flask==3.0.0
pandas==2.1.4
requests==2.31.0
ebaysdk==2.2.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
Pillow==12.0.0
openpyxl==3.1.5
